Time: 108 ms (5.59%), Space: 30 MB (61.54%) - LeetHub

+++ 1354-construct-target-array-with-multiple-sums/1354-construct-target-array-with-multiple-sums.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPossible(vector<int>& target) {
-        unsigned int sum = 0;
+        long int sum = 0;
         priority_queue<int> q;
         for (int i=0;i<target.size();i++) 
             sum+=target[i], q.push(target[i]);
